+++ kernel/module/main.c
@@ -231,7 +231,7 @@ static int find_any_unique_sec(const struct load_info *info, const char *name)
 static void *section_addr(const struct load_info *info, const char *name)
 {
 	/* Section 0 has sh_addr 0. */
-	return (void *)info->sechdrs[find_sec(info, name)].sh_addr;
+	return (void *)(ulong)info->sechdrs[find_sec(info, name)].sh_addr;
 }
 
 /* Find a module section, or NULL.  Fill in number of "objects" in section. */
@@ -244,7 +244,7 @@ static void *section_objs(const struct load_info *info,
 
 	/* Section 0 has sh_addr 0 and sh_size 0. */
 	*num = info->sechdrs[sec].sh_size / object_size;
-	return (void *)info->sechdrs[sec].sh_addr;
+	return (void *)(ulong)info->sechdrs[sec].sh_addr;
 }
 
 /* Find a module section: 0 means not found. Ignores SHF_ALLOC flag. */
@@ -273,7 +273,7 @@ static __maybe_unused void *any_section_objs(const struct load_info *info,
 
 	/* Section 0 has sh_addr 0 and sh_size 0. */
 	*num = info->sechdrs[sec].sh_size / object_size;
-	return (void *)info->sechdrs[sec].sh_addr;
+	return (void *)(ulong)info->sechdrs[sec].sh_addr;
 }
 
 #ifndef CONFIG_MODVERSIONS
@@ -1444,7 +1444,7 @@ static bool ignore_undef_symbol(Elf_Half emachine, const char *name)
 static int simplify_symbols(struct module *mod, const struct load_info *info)
 {
 	Elf_Shdr *symsec = &info->sechdrs[info->index.sym];
-	Elf_Sym *sym = (void *)symsec->sh_addr;
+	Elf_Sym *sym = (void *)(ulong)symsec->sh_addr;
 	unsigned long secbase;
 	unsigned int i;
 	int ret = 0;
@@ -2621,7 +2621,7 @@ static int move_module(struct module *mod, struct load_info *info)
 				ret = -ENOEXEC;
 				goto out_err;
 			}
-			memcpy(dest, (void *)shdr->sh_addr, shdr->sh_size);
+			memcpy(dest, (void *)(ulong)shdr->sh_addr, shdr->sh_size);
 		}
 		/*
 		 * Update the userspace copy's ELF section address to point to
@@ -2759,7 +2759,7 @@ static struct module *layout_and_allocate(struct load_info *info, int flags)
 		return ERR_PTR(err);
 
 	/* Module has been copied to its final place now: return it. */
-	mod = (void *)info->sechdrs[info->index.mod].sh_addr;
+	mod = (void *)(ulong)info->sechdrs[info->index.mod].sh_addr;
 	kmemleak_load_module(mod, info);
 	codetag_module_replaced(info->mod, mod);
 
@@ -2797,7 +2797,7 @@ static int post_relocation(struct module *mod, const struct load_info *info)
 	sort_extable(mod->extable, mod->extable + mod->num_exentries);
 
 	/* Copy relocated percpu area over. */
-	percpu_modcopy(mod, (void *)info->sechdrs[info->index.pcpu].sh_addr,
+	percpu_modcopy(mod, (void *)(ulong)info->sechdrs[info->index.pcpu].sh_addr,
 		       info->sechdrs[info->index.pcpu].sh_size);
 
 	/* Setup kallsyms-specific fields. */
