+++ kernel/module/sysfs.c
@@ -195,7 +195,7 @@ static int add_notes_attrs(struct module *mod, const struct load_info *info)
 			nattr->attr.name = mod->sect_attrs->attrs[loaded].battr.attr.name;
 			nattr->attr.mode = 0444;
 			nattr->size = info->sechdrs[i].sh_size;
-			nattr->private = (void *)info->sechdrs[i].sh_addr;
+			nattr->private = (void *)(ulong)info->sechdrs[i].sh_addr;
 			nattr->read = sysfs_bin_attr_simple_read;
 			++nattr;
 		}
